Validating multiple views per model
Instead of returning the attributes bound to the view, this change returns all of the attributes on the model that have validation defined. When option argument is undefined this assume all properties on the model that have validation defined what to be validated. If option is defined then the specific attributes will be defined from that argument.